src/problemspace/misc/synonyms/requirements.txt
click==8.1.3
gensim==3.8.3
joblib==1.2.0
nltk==3.5
numpy==1.23.4
PyYAML==5.4.1
regex==2022.10.31
scipy==1.9.3
smart-open==5.2.1
tqdm==4.64.1
